+++ src/dynamite/states.py
@@ -213,11 +213,7 @@ class State:
             A representation of the state.
         """
 
-        s2i_result = self.subspace.state_to_idx(self.str_to_state(s, self.L))
-        if isinstance(s2i_result, tuple):
-            idx, sign = s2i_result
-        else:
-            idx = s2i_result
+        idx = self.subspace.state_to_idx(self.str_to_state(s, self.L))
 
         if idx == -1:
             raise ValueError('Provided initial state not in requested subspace.')